requirements.txt
alembic==1.13.1
banal==1.0.6
certifi==2024.6.2
charset-normalizer==3.3.2
dataset==1.6.2
diskcache==5.6.3
greenlet==3.0.3
idna==3.7
json-flatten==0.3
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==1.4.52
typing_extensions==4.12.2
urllib3==2.2.2

Flask~=3.0.3